Use lighter grey for group authority on sidebar
Previously the color of the authority next to the logo on the
sidebar was a dark grey like the rest of the text on the sidebar.
This will give it a lighter color and was added into the _base
as organization color (since in the future this prob won't be
just authority but actualy organization).